refactor: split up key pair implementation
