+++ packages/client-core/src/key_pair/index.ts
@@ -1,3 +1,5 @@
 export { KeyType } from './constants';
-export { KeyPair, KeyPairEd25519, Signature } from './key_pair';
+export { KeyPair } from './key_pair';
+export { Signature } from './key_pair_base';
+export { KeyPairEd25519 } from './key_pair_ed25519';
 export { PublicKey } from './public_key';
