+++ packages/client-core/src/key_pair/key_pair.ts
@@ -1,20 +1,7 @@
-import { baseEncode, baseDecode } from 'borsh';
-import nacl from 'tweetnacl';
+import { KeyPairBase } from './key_pair_base';
+import { KeyPairEd25519 } from './key_pair_ed25519';
 
-import { KeyType } from './constants';
-import { PublicKey } from './public_key';
-
-export interface Signature {
-    signature: Uint8Array;
-    publicKey: PublicKey;
-}
-
-export abstract class KeyPair {
-    abstract sign(message: Uint8Array): Signature;
-    abstract verify(message: Uint8Array, signature: Uint8Array): boolean;
-    abstract toString(): string;
-    abstract getPublicKey(): PublicKey;
-
+export abstract class KeyPair extends KeyPairBase {
     /**
      * @param curve Name of elliptical curve, case-insensitive
      * @returns Random KeyPair based on the curve
@@ -38,58 +25,5 @@ export abstract class KeyPair {
         } else {
             throw new Error('Invalid encoded key format, must be <curve>:<encoded key>');
         }
-    }
-}
-
-/**
- * This class provides key pair functionality for Ed25519 curve:
- * generating key pairs, encoding key pairs, signing and verifying.
- */
-export class KeyPairEd25519 extends KeyPair {
-    readonly publicKey: PublicKey;
-    readonly secretKey: string;
-
-    /**
-     * Construct an instance of key pair given a secret key.
-     * It's generally assumed that these are encoded in base58.
-     * @param {string} secretKey
-     */
-    constructor(secretKey: string) {
-        super();
-        const keyPair = nacl.sign.keyPair.fromSecretKey(baseDecode(secretKey));
-        this.publicKey = new PublicKey({ keyType: KeyType.ED25519, data: keyPair.publicKey });
-        this.secretKey = secretKey;
-    }
-
-    /**
-     * Generate a new random keypair.
-     * @example
-     * const keyRandom = KeyPair.fromRandom();
-     * keyRandom.publicKey
-     * // returns [PUBLIC_KEY]
-     *
-     * keyRandom.secretKey
-     * // returns [SECRET_KEY]
-     */
-    static fromRandom() {
-        const newKeyPair = nacl.sign.keyPair();
-        return new KeyPairEd25519(baseEncode(newKeyPair.secretKey));
-    }
-
-    sign(message: Uint8Array): Signature {
-        const signature = nacl.sign.detached(message, baseDecode(this.secretKey));
-        return { signature, publicKey: this.publicKey };
-    }
-
-    verify(message: Uint8Array, signature: Uint8Array): boolean {
-        return this.publicKey.verify(message, signature);
-    }
-
-    toString(): string {
-        return `ed25519:${this.secretKey}`;
-    }
-
-    getPublicKey(): PublicKey {
-        return this.publicKey;
     }
 }
